Rename the base path, update README.md with better docs
