+++ README.md
@@ -1,2 +1,20 @@
-# operator-libs-linux
-Linux helper libraries for the operator framework
+# Operator Linux Libs (operator-libs-linux)
+
+## Description
+
+operator-libs-linux is a set of [charm libraries] which can be used for managing and manipulating
+Debian packages, Snap packages, system repositories, users and groups, and other operations needed
+for machine charm usage on Linux.
+
+## Usage
+
+Each library contains information on usage and code examples. They are meant to be complete as
+standalone libraries, and should be managed as [charm libraries], with installation via `fetch-lib`,
+after which they may be imported and used as normal charms.
+
+## Contributing
+
+Please see the [Juju SDK docs](https://juju.is/docs/sdk) for guidelines on enhancements to this
+charm following best practice guidelines, and `CONTRIBUTING.md` for developer guidance.
+
+[charm libraries]: https://juju.is/docs/sdk/libraries
