+++ lib/charm/operator_libs_linux/v0/snap.py
